Move search by profile url in connections. * Add authType property on connectionDto * remove jobTitle property. Use headline everywhere
